data-pipeline-lambda-functions/lambda/volvo-infleet/requirements.txt
boto3
anyio==4.4.0
aiohttp==3.10.3
certifi==2024.7.4
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.5
idna==3.7
JSON-log-formatter==1.0
numpy==2.0.1
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tenacity==8.5.0
tzdata==2024.1
urllib3==2.2.2